Move registry into gauge module
